Remove duplication from Esri and WFS reducers
