+++ src/reducers/wfs.js
@@ -21,39 +21,12 @@
 import uuid from 'uuid';
 
 import {WFS} from '../action-types';
+import {addSource, removeSource} from './source';
 
 const defaultState = {
   sources: {},
   actions: {},
 };
-
-/** Add a source to the state.
- * @param {Object} state Current state.
- * @param {Object} action Action to handle.
- *
- * @returns {Object} The new state.
- */
-function addSource(state, action) {
-  const new_source = {};
-  new_source[action.sourceName] = action.sourceDef;
-
-  const new_sources = Object.assign({}, state.sources, new_source);
-  return Object.assign({}, state, {
-    sources: new_sources
-  });
-}
-
-/** Remove a source from the state.
- * @param {Object} state Current state.
- * @param {Object} action Action to handle.
- *
- * @returns {Object} The new state.
- */
-function removeSource(state, action) {
-  const new_sources = Object.assign({}, state.sources);
-  delete new_sources[action.sourceName];
-  return Object.assign({}, state, {sources: new_sources});
-}
 
 /** Add a WFS action to the state.
  * @param {Object} state Current state.
